+++ query/query_executor.go
@@ -13,7 +13,7 @@ import (
 
 	"github.com/influxdata/influxdb/models"
 	"github.com/influxdata/influxql"
-	"github.com/uber-go/zap"
+	"go.uber.org/zap"
 )
 
 var (
@@ -141,9 +141,6 @@ type ExecutionContext struct {
 	// Output channel where results and errors should be sent.
 	Results chan *Result
 
-	// Hold the query executor's logger.
-	Log zap.Logger
-
 	// A channel that is closed when the query is interrupted.
 	InterruptCh <-chan struct{}
 
@@ -219,7 +216,7 @@ type QueryExecutor struct {
 
 	// Logger to use for all logging.
 	// Defaults to discarding all log output.
-	Logger zap.Logger
+	Logger *zap.Logger
 
 	// expvar-based stats.
 	stats *QueryStatistics
@@ -229,7 +226,7 @@ type QueryExecutor struct {
 func NewQueryExecutor() *QueryExecutor {
 	return &QueryExecutor{
 		TaskManager: NewTaskManager(),
-		Logger:      zap.New(zap.NullEncoder()),
+		Logger:      zap.NewNop(),
 		stats:       &QueryStatistics{},
 	}
 }
@@ -265,7 +262,7 @@ func (e *QueryExecutor) Close() error {
 
 // SetLogOutput sets the writer to which all logs are written. It must not be
 // called after Open is called.
-func (e *QueryExecutor) WithLogger(log zap.Logger) {
+func (e *QueryExecutor) WithLogger(log *zap.Logger) {
 	e.Logger = log.With(zap.String("service", "query"))
 	e.TaskManager.Logger = e.Logger
 }
@@ -304,7 +301,6 @@ func (e *QueryExecutor) executeQuery(query *influxql.Query, opt ExecutionOptions
 		QueryID:          qid,
 		Query:            task,
 		Results:          results,
-		Log:              e.Logger,
 		InterruptCh:      task.closing,
 		ExecutionOptions: opt,
 	}