+++ services/storage/store.go
@@ -9,7 +9,7 @@ import (
 	"github.com/influxdata/influxdb/models"
 	"github.com/influxdata/influxdb/services/meta"
 	"github.com/influxdata/influxdb/tsdb"
-	"github.com/uber-go/zap"
+	"go.uber.org/zap"
 )
 
 type Store struct {
@@ -20,15 +20,15 @@ type Store struct {
 		ShardGroupsByTimeRange(database, policy string, min, max time.Time) (a []meta.ShardGroupInfo, err error)
 	}
 
-	Logger zap.Logger
+	Logger *zap.Logger
 }
 
 func NewStore() *Store {
-	return &Store{Logger: zap.New(zap.NullEncoder())}
+	return &Store{Logger: zap.NewNop()}
 }
 
 // WithLogger sets the logger for the service.
-func (s *Store) WithLogger(log zap.Logger) {
+func (s *Store) WithLogger(log *zap.Logger) {
 	s.Logger = log.With(zap.String("service", "store"))
 }