+++ services/udp/service.go
@@ -12,7 +12,7 @@ import (
 	"github.com/influxdata/influxdb/models"
 	"github.com/influxdata/influxdb/services/meta"
 	"github.com/influxdata/influxdb/tsdb"
-	"github.com/uber-go/zap"
+	"go.uber.org/zap"
 )
 
 const (
@@ -56,7 +56,7 @@ type Service struct {
 		CreateDatabase(name string) (*meta.DatabaseInfo, error)
 	}
 
-	Logger      zap.Logger
+	Logger      *zap.Logger
 	stats       *Statistics
 	defaultTags models.StatisticTags
 }
@@ -67,7 +67,7 @@ func NewService(c Config) *Service {
 	return &Service{
 		config:      d,
 		parserChan:  make(chan []byte, parserChanLen),
-		Logger:      zap.New(zap.NullEncoder()),
+		Logger:      zap.NewNop(),
 		stats:       &Statistics{},
 		defaultTags: models.StatisticTags{"bind": d.BindAddress},
 	}
@@ -300,7 +300,7 @@ func (s *Service) createInternalStorage() error {
 }
 
 // WithLogger sets the logger on the service.
-func (s *Service) WithLogger(log zap.Logger) {
+func (s *Service) WithLogger(log *zap.Logger) {
 	s.Logger = log.With(zap.String("service", "udp"))
 }